Introduce InstallSnapshotStreaming to conserve RAM
This commit does two things:

 1. It uses our knowledge of the snapshot size in order to allocate just
    as much memory as we need to read the entire snapshot in memory
    (before: 442M RSS, after: 174M RSS, both with a 161M snapshot).

 2. It does away with reading the entire snapshot into memory and rather
    lets net/http stream from disk (client) to disk (server). With the
    new code path, the measured RSS during an InstallSnapshotStreaming
    request is about 1M above the levels that can be observed during
    normal operation.

Why bother about point 1 when point 2 fixes the issue? So that upgrades
are as smooth as possible. There is a fallback so that you can run nodes
which support InstallSnapshotStreaming in a network with nodes that do
not support it yet. This makes rollouts and rollbacks easier.

fixes https://github.com/robustirc/robustirc/issues/130